Allow ability to disable self-registration
In our use case, we'd like not to allow people to register into our
portal, and would prefer to disable that option declaratively via
`values.yaml` file.

I'm not certain that this will be picked up by harbor application but
previously similar thing has been part of the AdminServer. Now that it's
gone there is no trace of using this variable.

Re-adding it and allowing a customization would be great.

Signed-off-by: Rafal Proszowski <rafal.proszowski@digital.cabinet-office.gov.uk>