Allow mounting non-self signed certs.
This is needed because usually when a certificate is not self-signed,
and generated by
[cert-manager](https://github.com/jetstack/cert-manager) the tls secret
will not contain a `ca.crt` field.
Without this, certificates generated by cert-manager can't be used with
harbor. It seems as though mounting `tls.crt` will work if the
certificate is rooted at a trusted CA, which cert-manager certs
typically are.

Signed-off-by: jose <jose@autonomic.ai>