corrects keys to use for notary tls secret
when using a custom notary.secretName secret to avoid re-generating
notary secrets on each template render; the docs said to populate the
secret with "tls.ca", "tls.crt" and "tls.key" key names... this was
incorrect as the template appears to expect "ca", "crt" and "key" key
names.

update the README to reflect this.