Skip unknown expose.type
This commit adjust the creation of the service of nginx.  It will skip
the creation of the service if the `expose.type` is set to an unknown
value.
One benefit of this change is that user can create the service to expose
Harbor first, and deploy other components using the chart.

Signed-off-by: Daniel Jiang <jiangd@vmware.com>