Generate password for internal redis
To harden the security for the internal redis service, a strong
redis password is generated if one is not explicitly configured
in values.yaml.

This requires several helm chart changes to ensure that the
password is not re-generated with every `helm upgrade`:
* a k8s secret is used to store not only the redis password, but
also all other redis connection URL variables that incorporate it
* the template logic gives preference to the redis password value
in values.yaml, if one is present. To avoid re-generating the
password during `helm upgrade`, the previous password value is
extracted from the existing secret, if present.
* an annotation is added to all deployments and stateful-sets that
reference the redis password or one of the connection URLs, to ensure
they are redeployed if the password changes. The fact that the
password is not re-generated with each `helm upgrade` reduces the
set of impacted workload objects that need to be updated during
an upgrade.
* the redis URL used for the jobservice component is removed
from the configmap and replaced with an environment variable
* remove the redis password from the registry config map. It's
already covered by an environment variable
* using a named template to generate the random password for redis
is not a good solution when the password is referenced from multiple
places, because the template generates a different password each time
it is invoked. For this reason, all redis related named templates
(e.g. `harbor.redis.urlForCore`, `harbor.redis.urlForJobservice`)
have been modified to expect the password to be supplied as a
`.password` parameter.

Signed-off-by: Stefan Nica <snica@suse.com>