Add values.yaml options to control TLS certificate rotation
When auto-generated TLS certificates are used for external and internal
TLS, these certificates will be re-generated with every helm run, which
can have undesired effects, such as:
* external TLS certificates being re-generated during upgrades, which
requires users to re-configure their clients to accept the new CA
certificate. This can currently be alleviated by making a copy of the
existing secret and providing that as a secret certSource, but it's
an unnecessary complication that can be avoided by extending the
helm chart templates to cover this case.
* "helm upgrade" is also used for Day 2 operations other than upgrades,
such as making targeted configuration changes (e.g. password changes
for the external database or redis services, resizing the persistent
storage volume for a particular component). It is expected that such
operations only restart the pods of the affected services, to reduce
downtime. However, when internal TLS is enabled with auto-generated
certificates, almost all internal services are restarted, leading
to unnecessary load and downtime.

This change introduces a `rotateCert` configuration option to all places
that involve auto-generating certificates:
* external TLS
* internal TLS
* notary communications (also internal TLS, but one that cannot be
disabled for technical reasons)
* core token encryption/decryption - currently using a TLS certificate
and private key that are hard-coded in the helm chart (which can be
labeled as a security risk)

The `rotateCert` configuration option may be set to `false` in the
helm chart configuration (or supplied as a helm `--set` command line
argument) to explicitly disable re-generating TLS certificates during
a "helm upgrade" run where such a change isn't required. When set to
`false', the existing TLS certificates are loaded from their respective
secrets and reused.

Fixes: #786

Signed-off-by: Stefan Nica <snica@suse.com>