fix(ingress): Reorder portal path for AWS ALBs
AWS Load Balancer rules are evaluated in priority order, highest to
lowest from top to bottom. This change moves the all-matching / prefix
rule to the bottom of the list to unmask the more specific core paths.
This allows use of an AWS Load Balancer with Harbor.

Previously, all requests, even those matching core paths like /v2, would
be routed to the portal service because the portal path was evaluated as
the highest priority rule.

Signed-off-by: Samantha Baldwin <sbaldwin@actbluetech.com>