feat: Allow specifying strategy for deployments
Add a `strategy` field to components with a scalable `Deployment`. In
case of many replicas one might want to have tighter control of how how
updates are rolled out.

Furthermore, we have the following use-case: We are running a HA
deployment with 2 registry replicas (no PVC) and an `updateStrategy.type`
of `Recreate`. Even though our registry replicas are stateless without
volumes, both of them are taken down at the same time when we apply
updates due to the `Recreate` option, which we need for the stateful
components.

This change would allow us to control the update strategy for each
component individually, if necessary.

Signed-off-by: Hauke Lange <hla@siteimprove.com>